requirements.txt
alembic==1.7.4
autopep8==1.6.0
bcolz-zipline==1.2.4
Bottleneck==1.3.2
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
empyrical-reloaded==0.5.8
fmpsdk==20211011.0
greenlet==1.1.2
h5py==3.5.0
idna==3.3
importlib-metadata==4.8.1
importlib-resources==5.3.0
intervaltree==3.1.0
iso3166==2.0.2
iso4217==1.6.20180829
Logbook==1.5.3
lru-dict==1.1.7
lxml==4.6.3
Mako==1.1.5
MarkupSafe==2.0.1
multipledispatch==0.6.0
multitasking==0.0.9
networkx==2.6.3
numexpr==2.7.3
numpy==1.21.3
pandas==1.2.5
pandas-datareader==0.10.0
patsy==0.5.2
pycodestyle==2.8.0
python-dateutil==2.8.2
python-dotenv==0.19.1
python-interface==1.6.1
pytz==2021.3
requests==2.26.0
scipy==1.7.1
six==1.16.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.26
statsmodels==0.13.0
TA-Lib==0.4.21
tables==3.6.1
toml==0.10.2
toolz==0.11.1
trading-calendars==2.1.1
urllib3==1.26.7
wrapt==1.13.2
yfinance==0.1.64
zipline-reloaded==2.1.1
zipp==3.6.0
